Arel::Nodes::Equality
