+++ lib/arel/extensions.rb
@@ -60,6 +60,7 @@ require 'arel/extensions/generate_series'
 require 'arel/extensions/rank'
 require 'arel/extensions/coalesce'
 require 'arel/extensions/not_equal'
+require 'arel/extensions/equality'
 
 require 'arel/extensions/visitors'
