+++ lib/arel/extensions/visitors.rb
@@ -6,25 +6,6 @@ module Arel
     class ToSql
       private
 
-      def visit_Arel_Nodes_Equality(o, collector)
-        right = o.right
-
-        collector = visit o.left, collector
-
-        case right
-        when Arel::Nodes::Unknown, Arel::Nodes::False, Arel::Nodes::True
-          collector << ' IS '
-          visit right, collector
-
-        when NilClass
-          collector << ' IS NULL'
-
-        else
-          collector << ' = '
-          visit right, collector
-        end
-      end
-
       def visit_Arel_Nodes_NamedFunction(o, collector)
         aggregate(o.name, o, collector)
       end
